% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{add_lines}
\alias{add_lines}
\title{Add veritcal and horizontal lines to plot}
\usage{
add_lines(
  fig,
  yintercepts = NULL,
  xintercepts = NULL,
  color = "black",
  opacity = 1
)
}
\arguments{
\item{fig}{a plotly plot}

\item{yintercepts}{y-axis intercept of horizontal lins}

\item{xintercepts}{x-axis intercept of vertical lines}

\item{color}{color of lines}

\item{opacity}{opacity of lines}
}
\value{
The formatted plotly plot
}
\description{
Add veritcal and horizontal lines to plot
}
