% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{add_titles}
\alias{add_titles}
\title{Add title to plotly plot}
\usage{
add_titles(fig, title = NULL, x.title = NULL, y.title = NULL, leg.title = NULL)
}
\arguments{
\item{fig}{a plotly plot}

\item{title}{plot title}

\item{x.title}{x-axis title}

\item{y.title}{x-axis title}
}
\value{
The formatted plotly plot
}
\description{
Add title to plotly plot
}
