% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{add_watermark}
\alias{add_watermark}
\title{Function to add watermark at centre of plot}
\usage{
add_watermark(fig, source, opacity, size = 1)
}
\arguments{
\item{fig}{a plotly plot}

\item{source}{source of image}

\item{opacity}{opacity of image}

\item{size}{size relative to plot}
}
\description{
Function to add watermark at centre of plot
}
