% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{bar_plotly}
\alias{bar_plotly}
\title{Create a vertically stacked bar chart}
\usage{
bar_plotly(
  plot_data,
  title = NULL,
  x.title = NULL,
  y.title = NULL,
  leg.title = NULL,
  colors = NULL,
  date_format = NULL
)
}
\arguments{
\item{title}{Figure title}

\item{x.title}{Title of x-axis}

\item{y.title}{Title of y-axis}

\item{leg.title}{Title of the legend}

\item{date_format}{Format to apply to date columns}

\item{plot.data}{A data.frame containing plot data}

\item{fig}{A pre-existing plot to which the traces can be added}
}
\value{
A plotly plot
}
\description{
Create a vertically stacked bar chart
}
