% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{check_column_names}
\alias{check_column_names}
\title{Checks expected columns are in a table}
\usage{
check_column_names(tb, expected_names)
}
\arguments{
\item{tb}{A data.frame or tibble}

\item{expected_names}{Expected columns names within tb}
}
\description{
Raises an error if expected columns are absent.
}
