% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{clean_plotly}
\alias{clean_plotly}
\title{Apply default formatting to plot}
\usage{
clean_plotly(fig)
}
\arguments{
\item{fig}{a plotly plot}
}
\value{
The plotly plot with bar buttons removed
}
\description{
Apply default formatting to plot
}
