% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{factor_dates}
\alias{factor_dates}
\title{Format and factor columns containing date data}
\usage{
factor_dates(tb, date_format, min_cols = 1)
}
\arguments{
\item{tb}{A data.frame or tibble}

\item{date_format}{Format to use with date columns}

\item{min_cols}{Minimum number of columns to be found}
}
\value{
The data.frame or tibble with formatted columns
}
\description{
Format and factor columns containing date data
}
