% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{find_date_cols}
\alias{find_date_cols}
\title{Find name of columns containing date data}
\usage{
find_date_cols(tb, min_cols = 1)
}
\arguments{
\item{tb}{A data.frame or tibble}

\item{min_cols}{Minimum number of columns to be found}
}
\value{
Names of columns containing dates
}
\description{
Find name of columns containing date data
}
