% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{line_plotly}
\alias{line_plotly}
\title{Create a line plot using plotly}
\usage{
line_plotly(
  plot_data,
  title = NULL,
  x.title = NULL,
  y.title = NULL,
  leg.title = NULL,
  colors = NULL,
  date_format = NULL
)
}
\arguments{
\item{plot_data}{A data.frame containing plot data}

\item{title}{Figure title}

\item{x.title}{Title of x-axis}

\item{y.title}{Title of y-axis}

\item{leg.title}{Title of the legend}

\item{colors}{A named vector supplying colors}

\item{date_format}{Format to apply to date columns}
}
\value{
A plotly plot
}
\description{
Create a line plot using plotly
}
