% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{order_xy_axis}
\alias{order_xy_axis}
\title{Order x and y axis if they are factors}
\usage{
order_xy_axis(fig, plot_data)
}
\arguments{
\item{fig}{a plotly plot}

\item{plot_data}{The data from which the plot was created}
}
\value{
The formatted plotly plot
}
\description{
Order x and y axis if they are factors
}
