% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{line_plotly}
\alias{line_plotly}
\title{Create a line plot using plotly}
\usage{
line_plotly(
  plot.data,
  date_format = NULL,
  title = NULL,
  x.title = NULL,
  y.title = NULL,
  leg.title = NULL,
  fig = plotly::plot_ly()
)
}
\arguments{
\item{plot.data}{A data.frame containing plot data}

\item{date_format}{Format to apply to date columns}

\item{title}{Figure title}

\item{x.title}{Title of x-axis}

\item{y.title}{Title of y-axis}

\item{leg.title}{Title of the legend}

\item{fig}{A pre-existing plot to which the traces can be added}
}
\value{
A plotly plot
}
\description{
Create a line plot using plotly
}
