% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pie_plotly}
\alias{pie_plotly}
\title{Create a pie chart}
\usage{
pie_plotly(plot_data, hole = 0, title = NULL, leg.title = NULL, colors = NULL)
}
\arguments{
\item{title}{Figure title}

\item{leg.title}{Title of the legend}

\item{plot.data}{A data.frame containing plot data}
}
\value{
A plotly plot
}
\description{
Create a pie chart
}
